backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
# psycopg2-binary==2.9.9  # Commented out for Windows - install separately if needed
# For PostgreSQL: Install pre-built wheel or use: pip install psycopg2-binary --only-binary :all:
# For development: SQLite is used by default (no installation needed)
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.1
requests==2.31.0
docker==7.0.0
python-dateutil==2.8.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
